+++ hslua-packaging/src/HsLua/Packaging/Module.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-|
 Module      : HsLua.Packaging.Module
-Copyright   : © 2019-2024 Albert Krewinkel
+Copyright   : © 2019-2026 Albert Krewinkel
 License     : MIT
 Maintainer  : Albert Krewinkel <tarleb@hslua.org>
 Stability   : alpha
@@ -13,6 +13,12 @@ module HsLua.Packaging.Module
   ( -- * Documented module
     Module (..)
   , Field (..)
+    -- * Constructors
+  , deffield
+  , withType
+  , withDescription
+  , withValue
+    -- * Loading
   , registerModule
   , preloadModule
   , preloadModuleWithName
@@ -22,12 +28,45 @@ module HsLua.Packaging.Module
 where
 
 import Control.Monad (forM_)
+import Data.Text (Text)
 import HsLua.Core
 import HsLua.Marshalling (Pusher, pushAsTable, pushList, pushName, pushText)
 import HsLua.ObjectOrientation.Operation (Operation (..), metamethodName)
 import HsLua.Packaging.Documentation
 import HsLua.Packaging.Types
+import HsLua.Typing (TypeSpec, anyType)
+import qualified HsLua.Core.Utf8 as Utf8
 import qualified HsLua.Packaging.Function as Fun
+
+-- | Create a new module field.
+deffield :: Name -> Field e
+deffield name = Field
+  { fieldName = name
+  , fieldPushValue = return ()
+  , fieldDoc = FieldDoc
+      { fieldDocName = Utf8.toText $ fromName name
+      , fieldDocType = anyType
+      , fieldDocDescription = mempty
+      }
+  }
+
+-- | Set a specific type for a field.
+withType :: Field e -> TypeSpec -> Field e
+withType fld typespec =
+  let doc = fieldDoc fld
+  in fld { fieldDoc = doc { fieldDocType = typespec }}
+
+-- | Add a value pusher to a field.
+withValue :: Field e -> LuaE e () -> Field e
+withValue fld pusher = fld { fieldPushValue = pusher }
+
+-- | Add a textual description to a field.
+withDescription :: Field e -> Text -> Field e
+withDescription fld descr =
+  let doc = fieldDoc fld
+  in fld { fieldDoc = doc { fieldDocDescription = descr } }
+
+infixl 0 `withType`, `withValue`, `withDescription`
 
 -- | Create a new module (i.e., a Lua table).
 create :: LuaE e ()
@@ -93,9 +132,9 @@ pushModule mdl = do
 
   -- # Fields
   --
-  forM_ (moduleFields mdl) $ \field -> do
-    pushText (fieldName field)
-    fieldPushValue field
+  forM_ (moduleFields mdl) $ \fld -> do
+    pushName (fieldName fld)
+    fieldPushValue fld
     rawset (nth 3)
   case moduleOperations mdl of
     [] -> pure ()